data_integrity_auth_lab7/requirements.txt
flask
flask-jwt-extended
pyodbc
werkzeug
python-dotenv
